requirements.txt
django==3.0
pymysql
python-environ
django-rest-framework
django-ckeditor
django-db-logger
